search_engines.py
# search_engines.py

# Import the necessary libraries
from github import Github
from googlesearch import search
import tweepy


# GitHub search function
def search_github(query):
    # Set up the GitHub API client
    g = Github()

    # Search for repositories matching the query
    repositories = g.search_repositories(query)

    # Extract relevant information from the repositories
    results = []
    for repo in repositories:
        results.append({
            'source': 'GitHub',
            'title': repo.full_name,
            'link': repo.html_url
        })

    return results

# HuggingFace search function
def search_huggingface(query):
    # Search for models on the Hugging Face Model Hub
    # Implement your own logic here using the Hugging Face API or other available methods

    results = []  # Placeholder for results

    return results

# Twitter search function
def search_twitter(query):
    # Set up the Twitter API client
    # Implement your own logic here using the Tweepy library or other available methods

    results = []  # Placeholder for results

    return results

# Google search function
def search_google(query):
    # Perform a Google search for the query
    search_results = search(query, num_results=10)

    # Extract relevant information from the search results
    results = []
    for result in search_results:
        results.append({
            'source': 'Google',
            'title': result['title'],
            'link': result['link']
        })

    return results
