requirements.txt
tabulate==0.8.9
github3.py==1.3.0
requests==2.26.0
tweepy==4.0.1
google==3.0.0
PyGithub==1.55
google==3.0.0
tweepy==4.0.1